chore: move unit tests to correct place for mssql and postgres
